sanad_system/requirements-full.txt
# Core Django
Django>=4.2.0,<5.0.0
Pillow>=10.0.0
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.17.0

# OCR (if needed)
pytesseract==0.3.10

# Database (SQLite is included with Python)

# Development Tools
python-dotenv>=0.19.0
whitenoise>=6.6.0

# ASGI Server (for production)
uvicorn>=0.22.0
uvloop>=0.17.0

# Other utilities
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system=="Windows"
